+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Move ClientProviderInterface into Heptacom\OpenAuth\ClientProvider\Contract namespace and used Contract pattern
 * Move ClientInterface into Heptacom\OpenAuth\Client\Contract namespace and used Contract pattern
 * Add RedirectBehaviour class to control redirect process
-* Move TokenPairFactory into Heptacom\OpenAuth\Struct namespace
+* Move TokenPairFactory into Heptacom\OpenAuth\Token\Contract namespace and used Contract pattern
 * Move TokenPairStruct into Heptacom\OpenAuth\Struct namespace and removed Shopware dependency
 * Move UserStruct into Heptacom\OpenAuth\Struct namespace and removed Shopware dependency